Cleaned up caching headers so that IE8 can cache images (Issue 157)
Forcibly remove caching and last-modified headers on rewritten HTML files,
as we have cache-extended all the resources on the HTML files, so it's
dangerous to cache the HTML files.  They should be rewritten each time
to give a chance for changed origin resources to be rewritten in.

Fixed race-condition in async serf fetching that caused rare crashes
and server-load issues with URLs that time out.  This primarily happened
with the browser-proxy pointing to the Apache server and the mod_pagespeed
slurp mode turned on, but the URL being requested was actually on the
Apache server.

Add several new statistics for measuring cache activity, rewriting overhead,
and option-parsing overhead (for .htaccess).

Add locking of resources during rewriting, so that multiple processes don't
try to optimize the same css, js, or images at the same time.

Add more system-testing for caching headers.