Changed the mod_rewrite bypass to be more robust.
I no longer depend on string literals in mod_rewrite.c but have
taken matters into my own hands by squirreling away the original
URL in request->notes, which provides me a string-map.

So now I let mod_rewrite do whatever it wants, and instead examine
the original URL only once, before mod_rewrite.  Note that the user
can, with a bogus rewrite rule, completely screw up mod_pagespeed
now, because the the authorization module will work with the output
of mod_rewrite.  In fact this was happening in my test because I
was rewriting to a URL that did not start with a /.