Add tests for update frequency of combined and nested resources.
Tweak the code a bit to help avoid spurious tsan warnings. (morlovich)

Rewire how we handle rewrites in asynchronous case so that we have
a place to hook in the starting of nested jobs and can run after them.

Nest step will be to change image_rewriter_ in Context to be a
CssImageRewriterAsync, invoking it from Context::RewriteImages,
and having it feed rewrite contexts from other filters to
Context::RegisterNested. (morlovich)

Do more careful error-checking of libjpeg results.  Most of these shouldn't fail according to the docs (they should only do that if we temporarily stop providing input, which we don't do), but pagespeed was seeing related failures and there's no actual harm in coding this more conservatively. (jmaessen)