Do not allow GoogleUrl methods to CHECK-fail in opt-mode.
Fixed a bug where not finding a LeafPostion would send a negative number to StringPiece's length argument.

Rework them to instead return NULL if the URL is invalid or the method doesn't make sense for this URL type (Ex: PathSansQuery() for data URL).

Also reworked DCHECKs to LOG(DFATAL) so that we log ERRORs if we hit this in production (should never happen).