Still getting failures in blink running open-source tests under valgrind on ubiquity. (jmarantz)
Disable lazyload if the browser doesn't support inlining. Also, don't call ComputeCriticalImages if the user agent doesn't support inlining. (nikhilmadan)

Add an option to control the maximum size of an image in CSS that is converted to webp. (nikhilmadan)

Add new flush early tests to test.gyp. Missed this in the previous CL. (nikhilmadan)