Refactors cache-creation semantics and ownership into the ResourceManager.
In Apache, we can have distinct caches per VirtualHost
(ResourceManager), so all of the caches structures currently
owned and referenced in factories should really be held in the
ResourceManager instead.

This was earlier dealt with ad hoc with redundant pointers and
assertions and this has grown unmaintainable.

This CL enables the proper management of memcached connections, per vhost,
and independent from the file-caches.

I believe it also makes it easier to adopt property-cache capabilities
in mod_pagespeed properly, though that was not the original intent of
this change.