Hookup SPDY-specific configuration. Now <ModPagespeedIf spdy> and <ModPagespeedIf !spdy> actually work.
Also make it possible to turn on spdy-configuration with a header,
X-PSA-Optimize-For-SPDY, so that people who use something other than
mod_spdy to terminate SPDY may be able to use this, and so I can test
this without requiring hard-to-build tools like spdycat.

This is the code part of issue 450.