Stop throwing away extra attributes on <link> when inlining.
Previously: <link rel=stylesheet href=... title=foo> -> <style>...</style>
Now: We do not inline that style because it has unexpected attributes.

Other possibilities might be to transfer the extra attributes to the <style> tag, but I don't think we can do that for all of them (for example title seems to only be talked about for external CSS).

Exposed by testing for issue 523.