Clean up mechanism for punting resource-rewrites on unhealthy cache so we get a consistent view.
In most cases the canceling of RewriteContexts is done centrally in
RewriteDriver, deleting the contexts rather than adding them to the
queue.  The exception is CssFilter which holds onto a created
RewriteContext after attempting to initiate it, so in that case we
most proactively avoid adding the filter.

In all cases we unit-test the filters both for HTML-Rewrite & Fetch
with unhealthy caches.