Turn on critical image beaconing by default, on systems where it is supported (ie not PSS).
Fix knock-on bugs related to the treatment of missing critical image data.  We
experimented with various settings (including having no data mean "everything
critical" regardless of whether data was pending or not).  We are for now going
with the PSS behavior, where if critical image data just hasn't arrived yet we
assume that images are NOT critical.