Add support for embedded NULs in StringMultiMap keys.  This probably saves some runtime (not specifically measured) by reducing the number of .as_string() calls needed to make temps for use in Lookup.
It costs some memory by forcing an extra StringPiece object
for use as the key, while still keeping the GoogleString
backing-store for that string in the value.

This has one small interface change: name(index) returns
a StringPiece now rather than a const char*.

Also a new method RemoveAllFromSortedArray was added, along with a
speed test illustrating why it's better to use that than a
StringInsensitiveSet to implement RemoveAll.