Simplify RewriteDriver memory management and waiter wake up. Instead of having a bunch of bools with slightly different semantics and complicated logic to determine when to delete and when to ->Signal, just use a categorized refcount (e.g. active rewrites, detached rewrites, etc.), and check IsDone before and after changing it.
There is a bit of a semantic change: previously ->Cleanup would roughly set a 
"now it's safe to care about refcount 0" bit, now it's actually a deref,
so it's no longer sort-of idempotent.

There are also a few more checks; in particular you can no longer just 
call StartParse lots of times without doing a FinishParse, a few 
tests stumbled on that. ->ReleaseDriver also complains when its handed 
something that has non-zero refcounts, so most spots that used it are now 
calling ->Cleanup() instead