Change by vchudnov:
More detailed error codes from the scanline interface classes.

This change defines a ScanlineStatus class that is used to report the
success or error of various operations in the Scanline{Reader,Writer}Interface 
classes and of the CreateScanline{Reader,Writer} utilities. The old APIs are still 
present for backwards compatibility, though it would probably be a good idea in a 
follow-up changes to convert the clients to the new convention and get rid of the old APIs.