Add option to allow webp files to be reconstructed for any user-agent.
I made this an option because I believe that we believe that it's better

to serve origin resources for webp URLs to non-webp-capable browsers so
that link-sharing does not fail. The reason that this change is necessary
is to support sites that want to use proxy-caches that cannot work with
Vary:Accept, in which case they will not be able to downgrade responses
to old browsers.
    
Fixes Issue 848.
    
See also Issue 846 which, means that setting this option 'off' doesn't
currently work, since our http cache lookups currently do not
incorporate any bits indicating whether the client can support webp.  Thus
setting the default of the new option to 'on' for now.