Rework JavascriptCodeBlock to use explicit (rather than lazy) rewriting and checking for success.
In another CL I'm working on, I ran into problems when I added a comment at the end of the JS file because that might change the result of ProfitableToRewrite().

AFAICT, there's no advantage to this being lazily rewritten. Additionally, now that we explicitly rewrite, we can avoid all the mutable variables.