Framework for rooting in implicit_cache_ttl_ms from RewriteOptions into every ResponseHeaders object. Currently we have a set_implicit_cache_ttl_ms() method, but many callers do not use that setter and thus end up with default options.
Create HttpOptions structure (and future uses that want to access options from http/ directory).

Store a copy of HttpOptions in RequestContext. These contexts are a good way to get options into ResponseHeaders constructor in AsyncFetch.

Store a pointer to RewriteOptions in Resource. Resources also create ResponseHeaders and storing a version in here is useful.

I tried to pass in the correct custom options in each case, but in many places I could not access custom options. These are all marked with TODOs (or Notes explaining why default options are fine). A follow-up will pipe in custom options in those cases that need them and verify that the rest do not. Note: I believe that this does not regress anything. Just exposed the places where we are using default rather than custom options.

Also TODO is getting rid of the default ResponseHeaders constructor and forcing all construction to explicitly pass in options.