(From vchudnov)
GIF images: fix quirks mode; pad in scanline interface

This CL:

(a) tweaks the quirks mode as described in b/16539233, and

(b) causes the GIF reader returned by CreateScanlineReader() to pad GIF frames when they are smaller than the GIF image (or, in GIF terminology, to pad GIF images when they are smaller than the GIF screen).

The latter is done by means of a new MultipleFramePaddingReader class to implement the actual padding. This allows FrameToScanlineReaderAdapter to focus on simply converting the function calls between the two APIs.