Improve interaction between Responsive and Inline image filters.
Specifically, we now do the following:
1) If the highest resolution image is below the inline threshold (or some lower threshold), use that high-res image inline as the src (no srcset). This includes the case where the image had native dimensions == height & width attributes and was below the inline threshold.
2) Else, disable inlining and just add a normal srcset.

Notably here, we never mix inline images with responsive images. Technically, it is possible to mix the two (have a 1x version with data URL, but 2x without). However, it seems like this is (A) wasteful for the 2x browsers, (B) complicated to understand and (C) unlikely to be particularly common/much value to support.